Move HTTP listeners to a sub-namespace
