+++ phpunit/functional/Glpi/Http/LegacyItemtypeRouteListenerTest.php
@@ -36,7 +36,7 @@ namespace tests\units\Glpi\Http;
 
 use Glpi\Controller\GenericListController;
 use Glpi\Controller\DropdownFormController;
-use Glpi\Http\LegacyItemtypeRouteListener;
+use Glpi\Http\Listener\LegacyItemtypeRouteListener;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\RunInSeparateProcess;
 use PHPUnit\Framework\TestCase;
