+++ src/Glpi/Http/Listener/LegacyAssetsListener.php
@@ -32,8 +32,10 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Http;
+namespace Glpi\Http\Listener;
 
+use Glpi\Http\LegacyRouterTrait;
+use Glpi\Http\ListenersPriority;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
