+++ src/Glpi/Http/Listener/LegacyRouterListener.php
@@ -32,9 +32,11 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Http;
+namespace Glpi\Http\Listener;
 
 use Glpi\Controller\LegacyFileLoadController;
+use Glpi\Http\LegacyRouterTrait;
+use Glpi\Http\ListenersPriority;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
